Fixed Algorithm and Created Fuction.
I fixed the algorithm and created a function to make it simpler to use. There is still an bug. The closer to the sides of the image the detected minerals are, the more off the detected location is. This can be fixed with a simple scalar, but I need to find the root of the bug.